+++ app/file_storage.py
@@ -7,57 +7,8 @@ import os
 from os import path, getcwd, mkdir, rmdir, remove, replace, listdir
 import json
 import time as t
-
-def get_type(type):
-    type = type.lower()[1:]
-    if type in ['txt', 'pdf', 'iso']: return type
-    if type in ['doc', 'docx']: return 'word'
-    if type in ['bmp', 'jpg', 'jpeg', 'png', 'gif', 'cr2', 'ico', 'icon']: return 'image'
-    if type in ['ppt', 'pptx', 'pps', 'ppsx']: return 'ppt'
-    if type in ['avi', 'mov', 'qt', 'asf', 'rm', 'mp4', 'mkv']: return 'video'
-    if type in ['mp3', 'wav']: return 'music'
-    if type in ['xls', 'xlsx']: return 'excel'
-    if type in ['zip', 'rar', '7z', 'gz', 'z']: return 'compression'
-    return 'file'
+from function import *
     
-def secure_path(path):
-    return (path.find('..') == -1)
-
-    
-def normalsize(s):
-    nexts = {'B':'K', 'K':'M', 'M':'G', 'G':'T'}
-    ns = 'B'
-    while s >=1024:
-        s /= 1024
-        ns = nexts[ns]
-            
-    return "%.4f"%s+ns
-          
-def dirsearch(str):
-    size = 0
-    time = 0
-    if path.exists(str):
-        if path.isfile(str):
-            try:
-                return path.getsize(str), path.getmtime(str)
-            except OSError as err:
-                abort(405)
-        else:
-            try:
-                file_list = listdir(str)
-                for file in file_list:
-                    s, t = dirsearch(str + '/' + file)                  
-                    size += s
-                    time = max(time, t)
-                return size, time
-            except OSError as err:
-                abort(405)
-    return size,time 
-    
-def TimeStampToTime(timestamp):
-    timestruct = t.localtime(timestamp)
-    return t.strftime('%Y-%m-%d %H:%M:%S',timestruct)    
-        
 app = Flask('file_storage')
 app.secret_key = 'some_secret'
 
@@ -131,27 +82,11 @@ def logout():
 
 @app.route('/dir_index/<path:path_name>')
 def index(path_name):
-    cpath = getcwd() + '/' + path_name   
-    index_list = dirsearch(cpath)
-    fd_list = listdir(cpath)
-    index_list = []
-    idx = 1
-    for file in fd_list:
-        if path.isfile(cpath + '/' + file):
-            size = normalsize(path.getsize(cpath + '/' + file))
-            time = TimeStampToTime(path.getmtime(cpath + '/' + file))
-            name, type = path.splitext(file)
-            index_list.append([name, type[1:], get_type(type), idx, size, time])
-        else:
-            size, time = dirsearch(cpath+ '/' + file)
-            size = normalsize(size)
-            if time == 0 : time = '-------------------------'
-            else: time = TimeStampToTime(time)
-            index_list.append([file, 'dir', 'folder', idx, size, time])
-            
-        idx += 1
+    while path_name[-1]=='/': path_name = path_name[:-1]
+    dpath = getcwd() + '/' + path_name   
+    index_list, size, time = dirsearch(dpath, path_name)
     return render_template('index.html', list = index_list, path = path_name)
-    
+
 @app.route('/download/<path:file_path>')
 def download(file_path):
     if not secure_path(file_path): abort(405)
@@ -165,19 +100,10 @@ def download(file_path):
     abort(404)
 
 @app.route('/upload/<path:dir_path>', methods=['POST'])    
-def upload(dir_path):
-    validtype = ['txt', 'pdf', 'doc', 'docx', 'bmp', 'jpg', 'jpeg', 'png', 'gif', 'cr2',
-             'ppt', 'pptx', 'pps', 'ppsx', 'avi', 'mov', 'qt', 'asf', 'rm', 'mp4', 'mkv',
-             'mp3', 'wav', 'xls', 'xlsx', 'zip', 'rar', '7z', 'iso', 'gz', 'z', 
-             'ico', 'icon', 'html', 'css', 'js', 'py', 'c', 'cpp' ,'java']
-             
+def upload(dir_path):             
     if not secure_path(dir_path): abort(405)
     f = request.files['file']
-    
-    def valid_file(name):
-        ext = name.rsplit('.',1)[1].lower()
-        return (ext in validtype) and (name.find('\\')==-1 and name.find('/')==-1)
-        
+            
     if f and valid_file(f.filename):
         while (dir_path[-1] == '/'): dir_path = dir_path[:-1]
         if not path.exists(dir_path):
@@ -215,30 +141,7 @@ def createfolder(dir_path):
     return json.dumps({'code':200, 'dn':dirname})
     
 @app.route('/delete/<path:file_path>', methods=["DELETE"])
-def delete(file_path):
-    def del_rec(str):
-        if path.exists(str):
-            if path.isfile(str):
-                try:
-                    remove(str)
-                except OSError as err:
-                    abort(405)
-            else:
-                try:
-                    file_list = listdir(str)
-                    for file in file_list:
-                        str1 = str + '/' + file
-                        del_rec(str1)                  
-                    rmdir(str)
-                except OSError as err:
-                    abort(405)
-                    
-    def parent(str):
-        k = len(str) - 1
-        while (k > 0) and (str[k] != '/'):
-            k = k - 1
-        return str[0:k]                
-        
+def delete(file_path):                
     if not secure_path(file_path): abort(405)
     while file_path[-1]=='/': file_path = file_path[:-1]
     root_dir = getcwd()
